landing-page-openai/requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Async support
aiofiles>=23.0.0

# Utility
rich>=13.0.0
colorama>=0.4.6
